Projects/ynab-enrichment/requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
beautifulsoup4==4.12.3
lxml==5.2.1
pytest==8.2.0
